Adjust references to loop_min.
When adding iterations to a loop, we change the value of loop_min of
that loop, so we now compensate for that by replacing existing
references to loop_min by (loop_min+iters). In practice, the source of
such references is the sliding window optimization for the inner loops
that cannot be aligned by adding iterations.